:recycle: ProductListViewModel 추상 클래스와 구현체 분리
