+++ app/src/main/res/layout/fragment_product_list.xml
@@ -112,7 +112,7 @@
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/product_latest_title"
-                    tools:context=".ui.productList.ProductListFragment"
+                    tools:context=".ui.productList.DefaultProductListFragment"
                     tools:itemCount="10"
                     tools:listitem="@layout/holder_product_history" />
